server/requirements.txt
langgraph 
langchain_google_genai 
python_dotenv 
langchain_community 
langchain_core 
fastapi
langchain_groq